requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# servidora/src/util/parser.py: 2
PyYAML == 6.0

# servidora/src/util/util.py: 3
fabric == 2.6.0

# servidora/src/pveapi/__init__.py: 1
# servidora/src/pveapi/power_op.py: 1
# servidora/src/pveapi/status.py: 1
# servidora/src/pveapi/vm.py: 1
proxmoxer == 1.2.0

# servidora/src/main.py: 1,2
pyTelegramBotAPI == 4.1.1

